feat: Refactorisation des tests present dans project entity, bon nommage de certains fichiers et module bien adapter
