+++ apps/server/src/application/dto/response/project-response.adapter.ts
